refactor: Added shapes folder
