+++ src/core/rt3-base.h
@@ -11,9 +11,9 @@
 #include "parser.h"
 #include "primitive.h"
 #include "surfel.h"
-#include "sphere.h"
+#include "../shapes/sphere.h"
 #include "../materials/material.h"
-#include "shape.h"
+#include "../shapes/shape.h"
 #include "primlist.h"
 #include "scene.h"
 #include "../integrators/integrator.h"
